add day 5 part 1 and part 2
part 2 took ~4mins to run with -02 so the brute force solution this
uses is not ideal.

some places online recommend either mapping input ranges to output ranges
or by implementing some short-circuits like if the end of the range is
strictly linear to the beginning, then the rest of the range can be skipped.

Pretty interesting.